+++ gzap/CHANGELOG.md
@@ -18,6 +18,8 @@ the apworld to contain your new tuning in order to test it.
   - Autodiscovery of addon apworlds containing gzdoom tuning/logic.
   - Logic for Eviternity and Eviternity II, by Akos.
   - Logic for Legacy of Rust, by RakeVuril.
+  - If you download the pk3 for your game from the web host, the client will
+    read the connection address from it and connect to AP automatically.
 - Fix:
   - Add `map07special` flag to Master Levels for Doom II MAP20, since it was
     originally designed to go in the MAP07 slot and relies on that behaviour.